app/requirements.txt
blinker==1.7.0
click==8.1.7
Faker==22.2.0
Flask==3.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
kafka-python==2.0.2
MarkupSafe==2.1.4
python-dateutil==2.8.2
six==1.16.0
Werkzeug==3.0.1
